Move stoplists to stoplist directory, re-rename stoplists
